% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_double_sorted_alpha.R
\name{calculate_alphas}
\alias{calculate_alphas}
\title{calculate alphas for double sorted portfolios}
\usage{
calculate_alphas(dt)
}
\arguments{
\item{dt}{data table with the return of a double sorted portfolio}
}
\value{
\code{dt} data table with the alphas for each double sorted portfolio
}
\description{
It calculates the alpha for the return of double sorted portfolio for each rank. The
rank are max ranks and factor rank(e.g small size for LMAX, MMAX and HMAX),  differcne between the highest and lowest rank (HMAX-LMAX) and average return
  for each factor controlling for max(e.g. average return for three portfolios of BIG,MED and SMA size for LMAX).
}
