% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fama_macbeth.R
\name{calculate_bivariate_avg}
\alias{calculate_bivariate_avg}
\title{calculate_bivariate_avg}
\usage{
calculate_bivariate_avg(dt, factors_names, names_displayed)
}
\arguments{
\item{dt}{data table with stock returns and ten factorS (e.g momentum,size,IV,max)}
}
\value{
\code{dt} data table with the average beta for each pair factors and
their p-value
}
\description{
It takes a data table with ten factors and returns
and run monthly regression for pair of factors on the
return  and take the beta average for each factor
and display it with the p-value
}
