% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fama_macbeth.R
\name{calculate_bivariate_avg_for_all_vars}
\alias{calculate_bivariate_avg_for_all_vars}
\title{calculate_bivariate_avg_for_all_vars}
\usage{
calculate_bivariate_avg_for_all_vars(dt, iv = FALSE)
}
\arguments{
\item{dt}{data table with stock returns and ten factorS (e.g momentum,size,IV,max)}
}
\value{
\code{dt} data table with the average beta for all factors and
their p-value
}
\description{
It takes a data table with ten factors and returns
and run monthly regression for each ten factors as indpedent and th
return is the dependent and take the beta average for each factor
and display them with the p-value
}
