% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_returns.R
\name{calculate_daily_returns}
\alias{calculate_daily_returns}
\title{ccalculate daily return}
\usage{
calculate_daily_returns(dt, country_code = NULL, save = FALSE,
  clean_NA = TRUE)
}
\arguments{
\item{dt}{data table with prices column}

\item{country_code}{the country in which the data belongs to.}

\item{save}{{optional} save data on working directory}
}
\value{
\code{dt} the orginal data table with additional column for daily returns
}
\description{
It takes a long formated data frame with prices, market value, book balue and volumn
and calculate the daily returns using log(price_today/price_yesterday)
}
\note{
the risk free rate still needs to be subracted from daily return
}
