% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_double_sorted_alpha.R
\name{calculate_double_sorted_alpha}
\alias{calculate_double_sorted_alpha}
\title{calculate double sorted alphas}
\usage{
calculate_double_sorted_alpha(dt)
}
\arguments{
\item{raw_stocks_info_file}{link of the stock information}
}
\value{
\code{dt} data table with double sorted alpha statstics for both
equally and value weighted portfolios
}
\description{
It takes a data table with return and ten other factors.
It calls get_factor_controlled_portfolio_stats function to calculates
 the alpha for a certain factor (e.g. size) of equally and double sorted
portfolios
}
