% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ff3.R
\name{calculate_ff3}
\alias{calculate_ff3}
\title{ccalculate daily fama-french}
\usage{
calculate_ff3(dt, monthly = F, country_code = NULL, save = FALSE)
}
\arguments{
\item{dt}{data table with prices column}

\item{country_code}{the country in which the data belongs to.}

\item{save}{{optioanl} save data}
}
\value{
\code{dt} the orginal data table with additional three columns for value, size
and ff3_por returns
}
\description{
It takes a long formated data frame with prices, market value, book balue and volumn
and calculate the daily SMB and HML.
It first calls for calculate_six_portfolios_returns to calculate the daily six portfolios
then make the SMB and HML calculation.
}
