% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_iv.R
\name{calculate_iv_e_returns}
\alias{calculate_iv_e_returns}
\title{calculate equally weighted returns for each iv rank}
\usage{
calculate_iv_e_returns(iv)
}
\arguments{
\item{iv}{data table with iv_rank and stock prices}
}
\value{
\code{dt} data table with portfolio return, rank and dates in yearmon format
}
\description{
It take a data table with IV of the past month and monthly return and
calculates iv return for each iv rank/class
}
