% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_iv.R
\name{calculate_iv_factor}
\alias{calculate_iv_factor}
\title{calculate returns for iv portfolio}
\usage{
calculate_iv_factor(iv, is_equally_weighted, portfolio = F, num_cuts)
}
\arguments{
\item{iv}{data table with iv rank and iv value}

\item{is_equally_weighted}{logical indicating wether the factor for equally or value weighted
returns}
}
\value{
\code{dt} the orginal data table with two more columns:  IV and iv_rank
}
\description{
It takes a long-formated data table with daily return, market premium, smb and hml
factors and calculates idiosyncratic volatility  and classify firms accordingly.
}
