% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_iv.R
\name{calculate_iv_factor_for_e_and_v_returns}
\alias{calculate_iv_factor_for_e_and_v_returns}
\title{calculate IV factor}
\usage{
calculate_iv_factor_for_e_and_v_returns(dt, portfolio = F, weighted = F,
  num_cuts)
}
\arguments{
\item{dt}{data table of the daily prices, market value and book-to-market
returns}
}
\value{
\code{dt} with iv_factor for both equally and value weighted returns
}
\description{
It takes a long-formated data table with daily prices, market premium, smb and hml
factors and call other function to calculates idiosyncratic volatility  and classify firms and
then it calculates the IV factor
}
