% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_iv.R
\name{calculate_iv_portfolio_alpha_and_raw_returns}
\alias{calculate_iv_portfolio_alpha_and_raw_returns}
\title{calculate alpha and raw return for portfolio}
\usage{
calculate_iv_portfolio_alpha_and_raw_returns(iv_ps, is_ew, dt, num_cuts)
}
\arguments{
\item{iv_ps}{data table with monthly return of each iv rank}

\item{is_ew}{logical, indicating wither the monthly return of each iv rank is equally-weighted
or value weighted.}

\item{dt}{data table with daily prices, market value and book to market. used for FF-3 calculation}
}
\value{
\code{dt} with alpha and raw returns for either equally-weighted or value-weighted iv portfolio
}
\description{
It takes a data table with the monthly return of each iv rank,
then it calculates the difference between the HMAX and LIV and merge them
with the dt of three-factor model. Next, it calls calculate_iv_portfolio_alpha_or_raw_returns function
twice to get the raw and alpha returns. It calls round_and_format function to format the alpha and raw
returns and then combine them in one data table
}
