% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_iv.R
\name{calculate_iv_portfolio_alpha_or_raw_returns}
\alias{calculate_iv_portfolio_alpha_or_raw_returns}
\title{calculates alpha or raw return for a iv portfolio}
\usage{
calculate_iv_portfolio_alpha_or_raw_returns(iv_ps, alphas = T, num_cuts)
}
\arguments{
\item{alphas}{logical, if True, it calculates the alpha return
returns}
}
\value{
\code{dt} with alpha or raw return for each iv rank
}
\description{
It takes a data table with monthly return of each iv rank and H-IV-LIV rank
and calculates either the alpha or raw raturn for each rank. It uses only the intercept and P-value
}
