% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_iv.R
\name{calculate_iv_rank}
\alias{calculate_iv_rank}
\title{calculate IV and classify firms accordingly}
\usage{
calculate_iv_rank(dt, portfolio_only = F)
}
\arguments{
\item{dt}{data table with daily prices,market value and book to market values}

\item{portfolio_only}{{logical} if True returns only iv for each stock}
}
\value{
\code{dt} the orginal data table with  IV value and iv_rank, monthly prices
and market vlaue
}
\description{
It takes a long-formated data table with daily return, market premium, smb and hml
factors and calculates idiosyncratic volatility  and classify firms accordingly.
}
