% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ff3.R
\name{calculate_market_excess_returns}
\alias{calculate_market_excess_returns}
\title{ccalculate daily market return}
\usage{
calculate_market_excess_returns(dt, monthly = F, country_code = NULL,
  save = FALSE)
}
\arguments{
\item{dt}{data table with weighted_return column}

\item{country_code}{the country in which the data belongs to.}

\item{save}{{optional} save data on working directory}
}
\value{
\code{dt} dates with market returns
}
\description{
It takes a long formated data frame with prices, market value, book balue and volumn
and calculate the daily weighted returns for all the stocks in the market
}
