% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_max.R
\name{calculate_max_e_returns}
\alias{calculate_max_e_returns}
\title{calculate equally-weighted return for stock sorted on max}
\usage{
calculate_max_e_returns(max)
}
\arguments{
\item{max}{data table with max rank and firms max}
}
\value{
\code{dt} with volum,prices,market value and book value
}
\description{
It takes a long-formated data table with daily returns  and calculates
 maxmim return for a firm in a given month and classifying firms accordingly by calling
 calculate_max function. Next, it calculates the  return for each
 max portfolio: low,med and high using ln(Pt1) - ln(Pt0)
}
