% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_max.R
\name{calculate_max_factor}
\alias{calculate_max_factor}
\title{calculate max factor for given portfolio}
\usage{
calculate_max_factor(max, is_equally_weighted, portfolio = F)
}
\arguments{
\item{max}{data table with max rank and max value}

\item{is_equally_weighted}{logical indicating wether the factor for equally or value weighted
returns}
}
\value{
\code{dt} the orginal data table with two more columns:  IV and iv_rank
}
\description{
It takes a long-formated data table with max rank and value and call one of two
functions (equally-weighted or value-weighted) to calculate reurn. Next, it calculates the
difference between high max and low max as the max factor.
}
