% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_max.R
\name{calculate_max_portfolio_alpha_or_raw_returns}
\alias{calculate_max_portfolio_alpha_or_raw_returns}
\title{calculates alpha or raw return for a max portfolio}
\usage{
calculate_max_portfolio_alpha_or_raw_returns(max_ps, alphas = T)
}
\arguments{
\item{alphas}{logical, if True, it calculates the alpha return
returns}
}
\value{
\code{dt} with alpha or raw return for each max rank
}
\description{
It takes a data table with monthly return of each max rank and H-MAX-LMAX rank
and calculates either the alpha or raw raturn for each rank. It uses only the intercept and P-value
}
