% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_max.R
\name{calculate_max_v_returns}
\alias{calculate_max_v_returns}
\title{calculate  value-weighted return for stock sorted on max}
\usage{
calculate_max_v_returns(max)
}
\arguments{
\item{max}{data table with max rank and firms max}
}
\value{
\code{dt} with volum,prices,market value and book value
}
\description{
It takes a long-formated data table with daily returns  and calculates
 maxmim return for a firm in a given month and classifying firms accordingly by calling
 calculate_max function. Next, it calculates the value-weighted return for each
 max portfolio: low,med and high. The weight for each portfolio is the total market
 cap for all of its stock divided by the market cap for the total market cap of all the
 three portfolios. The return is then normalised using natural log (1+return)
}
