% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_momentum.R
\name{calculate_momentum_e_rt}
\alias{calculate_momentum_e_rt}
\title{calculate returns for momentum portfolios}
\usage{
calculate_momentum_e_rt(mom)
}
\arguments{
\item{dt}{data table with monthly prices and momentum rank}
}
\value{
\code{dt} origianl data table with portfolio return (momentum portfolios)
}
\description{
calculate equally weighted-return for momentum portfolio
}
