% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_momentum.R
\name{calculate_momentum_factor}
\alias{calculate_momentum_factor}
\title{calculate momentum factor}
\usage{
calculate_momentum_factor(dt)
}
\arguments{
\item{dt}{data table with monthly prices and size columns}
}
\value{
\code{dt} with monthly return, firms, max rank and max value
}
\description{
calculate  momentum factor is the average difference between winners and
losers portfolios.
}
