% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bm_classfication.R
\name{calculate_monthly_bm_rank}
\alias{calculate_monthly_bm_rank}
\title{rank firms based on book-to-market (BM)}
\usage{
calculate_monthly_bm_rank(dt)
}
\arguments{
\item{dt}{data table with daily BM}
}
\value{
\code{dt} data table with monthly BM rank
}
\description{
It take a data table with daily BM. It keeps the BM of the last
day of each month for each firm and then lag them by six and then classify firms according
to their BM
}
