% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closing_prices_lag_one.R
\name{calculate_monthly_cl_prices_rank}
\alias{calculate_monthly_cl_prices_rank}
\title{rank firms based on closing price of the past month}
\usage{
calculate_monthly_cl_prices_rank(dt, num_cuts, double_sorted = T)
}
\arguments{
\item{dt}{data table with daily prices}
}
\value{
\code{dt} odata table with monthly prices rank
}
\description{
It take a data table with daily prices. It keeps the prices of the last
day of each month for each firm, lag them by one, and then classify firms according
}
