% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Illiquidity.R
\name{calculate_monthly_illiq_rank}
\alias{calculate_monthly_illiq_rank}
\title{calculate illiquidty}
\usage{
calculate_monthly_illiq_rank(dt, num_cuts, double_sorted = F)
}
\arguments{
\item{dt}{data table with daily prices}
}
\value{
\code{dt} with monthly illiquidty  value and rank
}
\description{
It takes a long-formated data table with daily prices. It calculates its return
and illiquidty and classify firms accordingly.
}
