% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_returns.R
\name{calculate_monthly_returns}
\alias{calculate_monthly_returns}
\title{ccalculate monthly return}
\usage{
calculate_monthly_returns(dt, country_code = NULL, save = FALSE)
}
\arguments{
\item{dt}{data table with prices column}

\item{country_code}{the country in which the data belongs to.}

\item{save}{{optional} save data on working directory}
}
\value{
\code{dt} the orginal data table with additional column for monthly returns
}
\description{
It takes a long formated data frame with prices, market value, book balue and volumn
and use the first day of the month per firm to divide it in the past month value minus 1
to get the simple monthly return. Next, it adds 1 to the return and take its log
}
