% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_reversal.R
\name{calculate_monthly_reversal}
\alias{calculate_monthly_reversal}
\title{rank firms based on the returns of the past month}
\usage{
calculate_monthly_reversal(dt, num_cuts, double_sorted = F)
}
\arguments{
\item{dt}{data table with daily prices}
}
\value{
\code{dt} odata table with monthly reversal value(past month return) and rank
}
\description{
It take a data table with daily prices. It calls calculate monthly returns
function, then lag the return by one and then classify firms accordingly
}
