% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_size_classfication.R
\name{calculate_monthly_size_rank}
\alias{calculate_monthly_size_rank}
\title{rank firms based on size}
\usage{
calculate_monthly_size_rank(dt, num_cuts, double_sorted = F)
}
\arguments{
\item{dt}{data table with daily market cap}
}
\value{
\code{dt} odata table with monthly market cap rank
}
\description{
It take a data table with daily market cap. It keeps the market value of the last
day of each month for each firm, takes the natural log, lag them by 1 and then classify them according
}
