% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_double_sorted_alpha.R
\name{calculate_portfolios_count_for_double_sorted_alpha}
\alias{calculate_portfolios_count_for_double_sorted_alpha}
\title{calculate number of portfolios for double sorted alpha}
\usage{
calculate_portfolios_count_for_double_sorted_alpha(dt, factor_rank_name,
  is_ew = T)
}
\arguments{
\item{dt}{data table with return and ten other factors}
}
\value{
\code{dt} data table with two coluls:one for date(year or yearmon) and the other
for the number of portfolios crospoding to the date
}
\description{
It calculate number of portfolios for a double sorted portfolio. For example,
a portfolio sorted on size and max.
}
