% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ff3.R
\name{calculate_six_portfolios_returns}
\alias{calculate_six_portfolios_returns}
\title{calculate fama-french six portfolios}
\usage{
calculate_six_portfolios_returns(dt, monthly = F)
}
\arguments{
\item{dt}{data table with daily returns, market cap, book-to-market}

\item{country_code}{{optional} the country in which the data belongs to.}

\item{save}{{optioanl} save data}
}
\value{
\code{dt} with the sum and return of fama-french six portflios
}
\description{
It starts with calling classify_firms_according_to_mv_and_bm function to make
the necessary classfication of firms. Next, it calculates the return for each one of the six
 portfolios: growth, neutral and value for big and small firms.
}
