% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_skewness.R
\name{calculate_skewness_rank}
\alias{calculate_skewness_rank}
\title{calculate iskewness and coskewness and classify firms accordingly}
\usage{
calculate_skewness_rank(dt, num_cuts, double_sorted = F)
}
\arguments{
\item{dt}{data table with prices, market value and book value}
}
\value{
\code{dt} t data table with coskewness and iskewness values and rank
}
\description{
It calculates iskewness and coskewness and then classify firms accordingly.
}
