% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ff3.R
\name{calculate_US_ff3}
\alias{calculate_US_ff3}
\title{calculate fama-french US values}
\usage{
calculate_US_ff3()
}
\value{
\code{dt} with three factor values with risk free rate
}
\description{
reterive, orginze the data from French website of the FF-3
}
