% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ff3.R
\name{classify_firms_according_to_mv_and_bm}
\alias{classify_firms_according_to_mv_and_bm}
\title{classfiys firms based on market cap and book-to-market}
\usage{
classify_firms_according_to_mv_and_bm(dt)
}
\arguments{
\item{dt}{data table with daily returns, market cap, book-to-market}

\item{mv_breakpoint}{{optional} market value break point to classify firms as big or small}
}
\value{
\code{dt} with the size and value columns
}
\description{
It classfiys firms according to their market cap. Firms Above the median are big
while firms below the median are small. Next, it classfys each size (big and small) to value,
neutral and growth. Firms with high book-to-market ratio are value while those with low
 book-to-marketa ratio are growth.
}
