% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_double_sorted_alpha.R
\name{construct_double_sorted_portfolios}
\alias{construct_double_sorted_portfolios}
\title{calculate double sorted alphas}
\usage{
construct_double_sorted_portfolios(dt, factor_dt, num_cuts, factor_rank_name,
  is_e_weighted = T)
}
\arguments{
\item{dt}{data table with return and ten other factors}
}
\value{
\code{dt} data table with double sorted alpha statstics for both
equally and value weighted portfolios
}
\description{
It takes a data table with return and ten other factors.
It calls get_factor_controlled_portfolio_stats function to calculates
 the alpha for a certain factor (e.g. size) of equally and double sorted
portfolios
}
