% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{export_results}
\alias{export_results}
\title{exporting the analysis results}
\usage{
export_results(country_code)
}
\arguments{
\item{country_code}{a string character indicating the country code}
}
\value{
export analysis results to excel
}
\description{
It exports the analysis results for a country to csv file
so that they can be used in writing
}
