% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_double_sorted_alpha.R
\name{format_table_output}
\alias{format_table_output}
\title{format alphas table}
\usage{
format_table_output(alphas_table, is_ew = T, num_cuts)
}
\arguments{
\item{alphas_table}{data table with alphas and p-vakues}

\item{is_ew}{logical, if TRUE, it changes the name of the factor column to EW(equially-weighted)}
}
\value{
\code{dt} data table with alpha and p value  in three columns
}
\description{
It take an alpha table with 81 columns
and change their format to long format to end up with three columns only
}
