% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_double_sorted_alpha.R
\name{get_avg_max_or_iv_protfolios}
\alias{get_avg_max_or_iv_protfolios}
\title{calculate the average returns for a factor}
\usage{
get_avg_max_or_iv_protfolios(data, factor_rank_name, num_cuts, is_max_p = T)
}
\arguments{
\item{data}{data table with the return of a double sorted portfolio}

\item{factor_rank_name}{factor name (e.g. Size_rank)}

\item{data}{data table with the return of a double sorted portfolio}
}
\value{
\code{dt} data table with the alphas for each double sorted portfolio
}
\description{
It calculates the average returns for a factor controlling on MAX (e.g. average return
for three portfolios of BIG,MED and SMA size for LMAX). In order to get the average return, it
controlls for max or iv but the factor values varis as illiustrated in the previous example.
}
