% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_double_sorted_alpha.R
\name{get_diff_max_protfolios}
\alias{get_diff_max_protfolios}
\title{calculate HMAX - LMAX returns}
\usage{
get_diff_max_protfolios(q_list, i, factor_rank_name)
}
\arguments{
\item{q_list}{a list with the return of q1, q2, q3 portfolios for
a given rank of the controlloed portfolio (e.g. Big or MED)}

\item{factor_rank_name}{factor name (e.g. Size_rank)}

\item{data}{data table with the return of a double sorted portfolio}
}
\value{
\code{dt} the orginal list bue with an extra column for the difference between HMAX and LMAX
}
\description{
It calculates the difference in returns between HMAX and LMAX portfolios.
}
