% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_summary.R
\name{get_double_sorted_alpha_by_country}
\alias{get_double_sorted_alpha_by_country}
\title{get double sort of max m min or IV by country}
\usage{
get_double_sorted_alpha_by_country(countries, num_cuts = 3,
  factor_type = "max")
}
\arguments{
\item{num_cuts}{numer of portfolios in each category
#'@param factory_type max, min or iv}

\item{country}{list of countries}
}
\value{
\code{dt} with double sort analysis result
}
\description{
results of double sort analysisIt of MAX , Min or IV.
}
