% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_ff3.R
\name{get_ew_daily_returns_and_ff3}
\alias{get_ew_daily_returns_and_ff3}
\title{merge equally weighted daily returns with the three factors}
\usage{
get_ew_daily_returns_and_ff3(dt)
}
\arguments{
\item{dt}{data table with prices, mar}

\item{country_code}{the country in which the data belongs to.}
}
\value{
\code{dt} with volum,prices,market value and book value
}
\description{
It takes a long-formated data table with prices, market value, book balue and volumn
 and calculates equally-weighted daily return and daily three factors and merge them.
}
