% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_summary.R
\name{get_fama_macbeth_bi_by_country}
\alias{get_fama_macbeth_bi_by_country}
\title{get fama macbeth bivariate factor by country}
\usage{
get_fama_macbeth_bi_by_country(countries)
}
\arguments{
\item{country}{list of countries}
}
\value{
\code{dt} with bivariate fama macbeth analysis result
}
\description{
results of fama macbeth  analysisIt o
}
