% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fama_macbeth.R
\name{get_fama_macbeth_univariate_values}
\alias{get_fama_macbeth_univariate_values}
\title{get_fama_macbeth_univariate_values}
\usage{
get_fama_macbeth_univariate_values(dt)
}
\arguments{
\item{dt}{data table with stock returns and ten factorS (e.g momentum,size,IV,max)}
}
\value{
\code{dt} data table with the average beta for each factor and p-value
}
\description{
It takes a data table with ten factors and returns
and call calculate univeriate avg function to get the beta average
of fama macbeth regression for each factor then group all
factors in one data table
}
