% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_iv.R
\name{get_iv_ew_and_vw_portfolio_returns}
\alias{get_iv_ew_and_vw_portfolio_returns}
\title{group iv alpha and raw returns for EW and VW portfolios}
\usage{
get_iv_ew_and_vw_portfolio_returns(dt, num_cuts)
}
\arguments{
\item{dt}{data table with daily  prices
returns}
}
\value{
\code{dt} with alphas and raw returns for equally-weighted and value-weighted portfolios
}
\description{
It takes a long-formated data table with daily return and calls
calculate_iv_factor_for_e_and_v_returns function to get the equally-weighted and
value-weighted returns for iv portfolio. Then, it calss calculate_iv_portfolio_alpha_and_raw_returns
function to get the alpha and raw returns for portfolio either equally-weighted or value-weighted. Finally,
it groups the equally-weighted and value-weighted alpha and raw returns in one data table
}
