% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_summary.R
\name{get_max_or_min_or_iv_by_country}
\alias{get_max_or_min_or_iv_by_country}
\title{get single sort of max , min or IV by country}
\usage{
get_max_or_min_or_iv_by_country(countries, num_cuts = 3,
  factor_type = "max")
}
\arguments{
\item{num_cuts}{numer of portfolios in each category}

\item{country}{list of countries}

\item{factory_type}{max, min or iv}
}
\value{
\code{dt} with single sort analysis result
}
\description{
results of single sort analysisIt of MAX , Min or IV.
}
