% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_double_sorted_alpha.R
\name{graph_double_sorted_portfolios_count}
\alias{graph_double_sorted_portfolios_count}
\title{graph count of portfolios for double sorted alpha portfolios}
\usage{
graph_double_sorted_portfolios_count(dt, is_ew = T, distrb = F)
}
\arguments{
\item{dt}{data table with return and ten other factors(e.g. size,BM,max)}

\item{is_ew}{logical, if FALSE graph the value weighted portfolio}

\item{distrb}{logical, if FALSE return the scater plot of portfolios count}
}
\value{
\code{dt} ggplot object
}
\description{
It calls the calculate_portfolios_count_for_double_sorted_alpha function
to calculate the number of portfolios then group them in one data table and graph
}
