% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{group_vars}
\alias{group_vars}
\title{group the ranking of various factors}
\usage{
group_vars(dt, n = 1, num_cuts = 3, country_code = FALSE, save = FALSE)
}
\arguments{
\item{dt}{data table with prices,mv,bm and volume}
}
\value{
\code{dt} data table with stocks nine factors and their ranking
}
\description{
It calls functions to calulate the nine factors (e.g. max,iv),then
merge them all in one data table
}
