% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{make_data_ready_for_modeling}
\alias{make_data_ready_for_modeling}
\title{make ready for analysis}
\usage{
make_data_ready_for_modeling(country_name = "saudi", n = 1, num_cuts = 3)
}
\value{
list with two data table
}
\description{
clean and make basic calculation to make ready for analaysis
}
