% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_double_sorted_alpha.R
\name{prepare_double_sorted_portfolios}
\alias{prepare_double_sorted_portfolios}
\title{prepare_double_sorted_portfolios}
\usage{
prepare_double_sorted_portfolios(dt)
}
\arguments{
\item{dt}{data table with daily prices,volume,mv and bm}
}
\value{
\code{dt} data table with monthly return and three factors model
}
\description{
It calculates the monthly return and construct the ff-3 model
to be ready to merge with controled variable(e.g size).
}
