% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roubst_se.R
\name{roubst_se}
\alias{roubst_se}
\title{calculates roubst standard error}
\usage{
roubst_se(mod)
}
\value{
\code{broom} with coefficent and p-value of the model
}
\description{
calculates roubst standard error and return coeficcents and P value
}
