% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_max.R
\name{round_and_format}
\alias{round_and_format}
\title{format the raw or alpha return results}
\usage{
round_and_format(avg_rt, alphas = T, num_cuts)
}
\arguments{
\item{avg_rt}{data table with alpha or raw return}

\item{alphas}{logical, if TRUE, it highlights that the return is for
the alpha of the three-factor model}
}
\value{
\code{dt} with max_factor for both equally and value weighted returns
}
\description{
It takes data table with the result of raw return or
alpha return for each max rank and round the return and p-value and format
the results.
max factors for for equally and value weighted portfolio
}
