% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_iv.R
\name{round_and_format_iv}
\alias{round_and_format_iv}
\title{format the raw or alpha return results}
\usage{
round_and_format_iv(avg_rt, alphas = T, num_cuts)
}
\arguments{
\item{avg_rt}{data table with alpha or raw return}

\item{alphas}{logical, if TRUE, it highlights that the return is for
the alpha of the three-factor model}
}
\value{
\code{dt} with iv_factor for both equally and value weighted returns
}
\description{
It takes data table with the result of raw return or
alpha return for each iv rank and round the return and p-value and format
the results.
iv factors for for equally and value weighted portfolio
}
