% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_max.R
\name{calculate_max_rank}
\alias{calculate_max_rank}
\title{calculate max and classify firms accordingly}
\usage{
calculate_max_rank(dt)
}
\arguments{
\item{dt}{data table with daily prices}
}
\value{
\code{dt} with monthly return, firms, max rank and max value
}
\description{
It takes a long-formated data table with daily prices, call for return
calculation and then calculates the maxmium return for each firm in each month and 
then classify firms accordingly.
}
