% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_double_sorted_alpha.R
\name{group_vars}
\alias{group_vars}
\title{group the ranking of various factors}
\usage{
group_vars(raw_stocks_info_file, country_code = FALSE, save = FALSE)
}
\arguments{
\item{raw_stocks_info_file}{link of the stock information}
}
\value{
\code{dt} data table with stocks nine factors and their ranking
}
\description{
It calls functions to calulate the nine factors (e.g. max,iv),then
merge them all in one data table
}
